// src/app/models/organizare.model.ts
export class Organizare {
    nr: number;
    id: number;
    judet: string;
    localitate: string;
    strada: string;
    nrStrada: string;
    mediu: string;
    nume: string;
    sirues: string;
    director: string;
    furnizorDate: string;
    nrFurnizor: string;
    explicatie: string;
    raspuns71: string;
    upload71: string;
    raspuns72: string;
    raspuns73: string;
    raspuns5: string;
    raspuns6: string;
    raspuns7: string;
    raspuns8: string;
    raspuns9: string;
    raspuns10: string;
    raspuns11: string;
    raspuns13: string;
    raspuns14: string;
    raspuns15: string;
    raspuns16: string;
    raspuns17: string;
    raspuns18: string;
    raspuns19: string;
    raspuns20: string;
    raspuns21: string;
    raspuns22: string;
    raspuns23: string;
    raspuns24: string;
    raspuns28: string;
    raspuns29: string;
    raspuns30: string;
    upload30: string;
    raspuns31: string;
    raspuns32: string;
    raspuns33: string;
    raspuns34: string;
    nrInregC: string;
    nrInreg: string;

    constructor(input?: any) {
        Object.assign(this, input);
    }
}